Add SimpleFile class
Since File is a built-in ruby class, SimpleFile was used instead.

This class is mainly used to calculate positional information for
tooling that need to report on the file name, line and column a
literal is found at. There is also a mechanism for adding locations
of newlines to support this.

It will probably become necessary to add getters for the filename
and source code.